fix complexity code smells
